Implemented authguard

+++ src/hotels/hotels.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadGatewayException,
+  BadRequestException,
+  Injectable,
+} from '@nestjs/common';
 import { BasicCrudService } from '../common/basic-crud.service';
 import { Hotel } from './entities/hotel.entity';
 import { EntityManager, FilterQuery } from '@mikro-orm/core';
@@ -6,7 +10,8 @@ import { HotelRepository } from './repositories/hotel.repository';
 import { CacheService } from '../cache/cache.service';
 import { OrganizationsService } from '../organizations/organizations.service';
 import { CreateHotelDto } from './dto/create-hotel.dto';
-import { FindHotelArgs } from './args/find-hotel.args';
+import { Roles } from '../users/enums/roles.enum';
+import { UserMeta } from '../auth/types/user-meta.type';
 
 @Injectable()
 export class HotelsService extends BasicCrudService<Hotel> {
@@ -19,10 +24,12 @@ export class HotelsService extends BasicCrudService<Hotel> {
     super(Hotel, hotelRepository, cacheService, entityManager);
   }
 
-  async createHotel(dto: CreateHotelDto) {
-    const organization = await this.organizationService.findOne(
-      dto.organizationId,
-    );
+  async createHotel(dto: CreateHotelDto, meta: UserMeta) {
+    if (meta.role !== Roles.ORGANIZATION) {
+      throw new BadGatewayException();
+    }
+
+    const organization = await this.organizationService.findOne(meta.id);
     if (!organization) {
       throw new Error('Organization not found');
     }
@@ -33,26 +40,40 @@ export class HotelsService extends BasicCrudService<Hotel> {
     if (existing) {
       throw new Error('Hotel already exists');
     }
-    return this.createOne(dto);
+    return this.createOne({ ...dto, organization });
   }
 
-  async removeHotel(id: number) {
-    const filter: FilterQuery<Hotel> = { id };
-    await this.deleteOne(filter);
-  }
-
-  async FindOneSafe(id: number) {
-    const filter: FilterQuery<Hotel> = { id };
-    return this.findOneOrFail(filter);
-  }
-
-  async findAll(args: FindHotelArgs) {
-    const filter: FilterQuery<Hotel> = {};
-    if (args.oganizationId) {
-      filter.organization = await this.organizationService.findOne(
-        args.oganizationId,
-      );
+  async removeHotel(id: number, meta) {
+    if (meta.role !== Roles.ORGANIZATION) {
+      throw new BadGatewayException();
     }
+
+    const filter: FilterQuery<Hotel> = { id };
+    const hotel = await this.findOneOrFail(filter);
+    const organization = await this.organizationService.findOneOrFail(meta.id);
+    if (hotel.organization === organization) {
+      return await this.deleteOne(filter);
+    } else throw new BadRequestException();
+  }
+
+  async FindOneSafe(id: number, meta: UserMeta) {
+    if (meta.role !== Roles.ORGANIZATION) {
+      throw new BadGatewayException();
+    }
+
+    const filter: FilterQuery<Hotel> = { id };
+    const hotel = await this.findOneOrFail(filter);
+    const organization = await this.organizationService.findOneOrFail(meta.id);
+    if (hotel.organization === organization) {
+      return hotel;
+    } else throw new BadRequestException();
+  }
+
+  async findAll(meta: UserMeta) {
+    const filter: FilterQuery<Hotel> = {};
+
+    filter.organization = await this.organizationService.findOne(meta.id);
+
     return this.findMany(filter);
   }
 }
