+++ src/rooms/rooms.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadGatewayException, Injectable } from '@nestjs/common';
 import { CacheService } from '../cache/cache.service';
 import { RoomRepository } from './repositories/room.repository';
 import { EntityManager, FilterQuery } from '@mikro-orm/core';
@@ -7,6 +7,8 @@ import { BasicCrudService } from '../common/basic-crud.service';
 import { Room } from './entities/room.entity';
 import { CreateRoomDto } from './dto/create-room.dto';
 import { FindRoomArgs } from './args/find-room.args';
+import { UserMeta } from '../auth/types/user-meta.type';
+import { Roles } from '../users/enums/roles.enum';
 
 @Injectable()
 export class RoomsService extends BasicCrudService<Room> {
@@ -19,7 +21,11 @@ export class RoomsService extends BasicCrudService<Room> {
     super(Room, roomRepository, cacheService, entityManager);
   }
 
-  async createRoom(dto: CreateRoomDto) {
+  async createRoom(dto: CreateRoomDto, meta: UserMeta) {
+    if (meta.role !== Roles.ORGANIZATION) {
+      throw new BadGatewayException()
+    }
+
     const hotel = await this.hotelsService.findOne(dto.hotelId);
     const existing = await this.roomRepository.findOne({
       roomNumber: dto.roomNumber,
@@ -31,7 +37,10 @@ export class RoomsService extends BasicCrudService<Room> {
     return this.createOne(dto);
   }
 
-  async removeRoom(id: number) {
+  async removeRoom(id: number, meta: UserMeta) {
+    if (meta.role !== Roles.ORGANIZATION) {
+      throw new BadGatewayException()
+    }
     const filter: FilterQuery<Room> = { id };
     await this.deleteOne(filter);
   }
